+++ docs/about/whats_new.rst
@@ -9,6 +9,7 @@ v1.9.next
 =========
 
 - Update and cross-reference 1.8 to 1.9 migration notes (:pull:`1686`)
+- Fix SimpleDocNav lineage handling with PostGIS index (:pull:`1687`)
 
 v1.9.0-rc12 (9th December 2024)
 ===============================